adding features for product crud and fix integration test for customer controller

+++ src/main/java/com/fiap/fast_food_tc/adapter/db/model/Customer.java
@@ -19,7 +19,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "customer_id", nullable = false, unique = true)
-    private Long customerId;
+    private Integer customerId;
 
     @Column(name = "document_number", nullable = false, unique = true, length = 11)
     private String documentNumber;
